Only run automation steps in the repo (#25)
In #23 we had our first external contribution and reviewbot and autosquash automation failed because of secrets access. We don't need either of these to run on external contributions, so limit running them to just the primary repo. Note that autosquash has to have its steps suppressed rather than the task because it's a required action (and needs to be a required action to function).