Add support for passing non-default arguments to uv pip compile (#55)
## Issue

There's been a lot of interest (see #46 and #53) in supporting custom arguments to be passed to `uv pip compile`, but this is complex because:
 - we didn't make it easy to override the defaults
 - we pass each argument explicitly from Bazel into the shell script
 - we have different scripts for compilation and testing

## Summary
Use templates to embed arguments directly rather than as bash variables. Refactor `_uv_pip_compile` to allow overriding the default arguments. Document defaults and update the README accordingly.