Support multiple input files and requirements-file references (#63)
## Issue
Fixes #58. Users of this module wish to use multiple input requirements files and refer to those files from a base `.in` file.

## Summary
Add support for passing additional files using a `data` argument. Add an example showing how this extra argument may be used to pass referenced `.in` files to generate a combined output `requirements.txt` file and document the new argument in the readme.