Transition uv to the correct platform (#122)
Currently, uv is selected for the exec platform. If it were used as part
of a build action, this would be correct, but as it is invoked by an
emitted shell script it is not. This causes it to fail when running
builds on one platform, but targeting another. This is the case, for
example, when running builds in a remote execution cluster with Linux
workers, but compiling for a Mac. This patch transitions uv to the
proper platform, allowing it to work in this situation.